# openapi.yaml
openapi: "3.0.3"
info:
  title: War Tiket Konser API
  description: |
    Sistem penjualan tiket konser dengan antrean virtual dan kursi terbatas.
    Kelompok 5 — Praktikum Microservices.
    
    **Masalah utama:** Satu kursi hanya boleh terjual satu kali walau seribu orang memintanya bersamaan.
    
    **Services:**
    - event-service (port 3001) — Kelola konser, jadwal, kategori kursi, dan harga
    - ticket-service (port 3002) — Kunci kursi sementara, konfirmasi, lepas kunci saat kedaluwarsa
    - payment-service (port 3003) — Terima pembayaran, konfirmasi, batalkan bila gagal/timeout
    - notification-service (port 3004) — Kirim e-ticket, pengingat, dan pemberitahuan gagal bayar
  version: "1.0.0"
  contact:
    name: Kelompok 5
servers:
  - url: http://localhost:3001
    description: event-service
  - url: http://localhost:3002
    description: ticket-service
  - url: http://localhost:3003
    description: payment-service
  - url: http://localhost:3004
    description: notification-service
...
paths:
...
  /orders:
    post:
      tags: [orders]
      summary: Buat pesanan — kunci kursi sementara (15 menit)
      description: |
        **ENDPOINT PALING KRITIS** — Ini adalah bottleneck utama sistem.
        Hanya satu orang yang bisa mengunci satu kursi.
        Menggunakan SELECT FOR UPDATE + Redis distributed lock untuk mencegah overselling.
        Kunci otomatis kedaluwarsa dalam 15 menit jika tidak dibayar.
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Kursi berhasil dikunci — lanjutkan ke pembayaran
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "409":
          description: Kursi sudah dikunci orang lain atau sudah habis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      tags: [orders]
      summary: Daftar pesanan milik pengguna
      operationId: getOrders
      parameters:
        - name: user_id
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
  /orders/{orderId}:
    get:
      tags: [orders]
      summary: Status satu pesanan
      operationId: getOrderById
      parameters:
        - $ref: "#/components/parameters/orderId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/NotFound"
  /orders/{orderId}/cancel:
    post:
      tags: [orders]
      summary: Batalkan pesanan dan lepas kunci kursi
      operationId: cancelOrder
      parameters:
        - $ref: "#/components/parameters/orderId"
      responses:
        "200":
          description: Pesanan dibatalkan dan kursi dilepas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/NotFound"
  /tickets/{ticketId}:
    get:
      tags: [tickets]
      summary: Status tiket yang sudah dikonfirmasi
      operationId: getTicketById
      parameters:
        - name: ticketId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          $ref: "#/components/responses/NotFound"

  # ══════════════════════════════════════════════════════
  # PAYMENT-SERVICE (port 3003)
  # ══════════════════════════════════════════════════════
...
components:
  parameters:
...
    orderId:
      name: orderId
      in: path
      required: true
      schema: { type: integer }
  responses:
    NotFound:
      description: Data tidak ditemukan
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Request tidak valid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Order:
      type: object
      properties:
        id:                { type: integer }
        user_id:           { type: string }
        event_id:          { type: integer }
        seat_category_id:  { type: integer }
        event_name:        { type: string }
        seat_category_name:{ type: string }
        price:             { type: number }
        status:
          type: string
          enum: [locked, confirmed, cancelled, expired]
        lock_expires_at:   { type: string, format: date-time }
        created_at:        { type: string, format: date-time }
    CreateOrderRequest:
      type: object
      required: [user_id, event_id, seat_category_id]
      properties:
        user_id:          { type: string, example: "user_105841117023" }
        event_id:         { type: integer, example: 1 }
        seat_category_id: { type: integer, example: 2 }
    Ticket:
      type: object
      properties:
        id:               { type: integer }
        order_id:         { type: integer }
        user_id:          { type: string }
        event_name:       { type: string }
        seat_category:    { type: string }
        seat_number:      { type: string }
        qr_code:          { type: string }
        status:
          type: string
          enum: [active, used, cancelled]
        issued_at:        { type: string, format: date-time }
...
    Error:
      type: object
      properties:
        error:   { type: string }
        message: { type: string }
        code:    { type: integer }